+++ ImageFeed/Helper/URLSession.swift
@@ -38,3 +38,4 @@ extension URLSession {
         return task
     }
 }
+
